minor debug

+++ foggie/ayan_acharyya_codes/run_foggie_sim.py
@@ -11,7 +11,7 @@
                  run run_foggie_sim.py --halo 2139 --level 1 --queue devel --automate --final_level 3 --dryrun
                  run run_foggie_sim.py --halo 2139 --level 1 --queue normal --automate --plot_projection --width 1000
                  run run_foggie_sim.py --halo 2139 --level 3 --queue long --gas --plot_projection --width 1000
-                 run run_foggie_sim.py --halo 8892 --queue long --forcedref --refsize 200 --reflevel 7
+                 run run_foggie_sim.py --halo 8892 --queue long --forcedref --refsize 200 --reflevel 7 --start_output RD0008
 
 """
 from header import *
@@ -409,7 +409,8 @@ def run_forcedref(ncores, nhours, args):
 
     # ------create new directory copy over required output-----------
     execute_command('mkdir -p ' + args.halo_dir + '/' + new_sim_name, args.dryrun)
-    execute_command('cp -r ' + args.halo_dir + '/' + args.sim_name + '/' + args.start_output + ' ' + args.halo_dir + '/' + new_sim_name + '/.', args.dryrun)
+    if not os.path.exists(args.halo_dir + '/' + new_sim_name + '/' + args.start_output):
+        execute_command('cp -r ' + args.halo_dir + '/' + args.sim_name + '/' + args.start_output + ' ' + args.halo_dir + '/' + new_sim_name + '/.', args.dryrun)
 
     # ---------make substitutions in RunScript file-----------------
     template_runscript = args.template_dir + '/RunScript_LX.sh'
